Factor benchmarking code into separate files
